# use server/util/proto library in cachetools