<!-- Optional: Provide additional context (beyond the PR title). -->

<!-- Optional: link a GitHub issue.
     Example: "Fixes #123" will auto-close #123 when the PR is merged. -->

We are using these methods to marshal, unmarshal repb.Tree, repb.Command, repb.Action and repb.Directory.
Benchmark for repb.Directory can be found in https://github.com/buildbuddy-io/buildbuddy/pull/5571

For other messages, the benchmark results are as follows:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 7900X 12-Core Processor            
                            │      Old      │                 New                  │
                            │    sec/op     │    sec/op     vs base                │
Marshal/pbName=Action-24       5.313µ ± 31%   4.324µ ±  9%        ~ (p=0.565 n=10)
Marshal/pbName=Tree-24         37.71m ±  1%   29.10m ±  2%  -22.83% (p=0.000 n=10)
Marshal/pbName=Command-24     13.461µ ±  8%   7.627µ ±  3%  -43.34% (p=0.000 n=10)
Unmarshal/pbName=Action-24     18.29µ ±  9%   12.58µ ± 14%  -31.26% (p=0.000 n=10)
Unmarshal/pbName=Tree-24       26.60m ±  1%   19.42m ±  1%  -26.99% (p=0.000 n=10)
Unmarshal/pbName=Command-24    45.67µ ±  8%   34.30µ ±  9%  -24.91% (p=0.000 n=10)
geomean                        197.8µ         141.5µ        -28.45%

                            │      Old      │                  New                  │
                            │      B/s      │      B/s       vs base                │
Marshal/pbName=Action-24      1.239Gi ± 41%   1.489Gi ±  8%        ~ (p=0.579 n=10)
Marshal/pbName=Tree-24        323.2Mi ±  1%   418.8Mi ±  2%  +29.59% (p=0.000 n=10)
Marshal/pbName=Command-24     320.0Mi ±  9%   564.8Mi ±  3%  +76.49% (p=0.000 n=10)
Unmarshal/pbName=Action-24    271.0Mi ±  9%   394.4Mi ± 17%  +45.56% (p=0.000 n=10)
Unmarshal/pbName=Tree-24      507.3Mi ±  1%   694.8Mi ±  1%  +36.97% (p=0.000 n=10)
Unmarshal/pbName=Command-24   238.7Mi ±  7%   317.9Mi ±  8%  +33.17% (p=0.000 n=10)
geomean                       403.4Mi         561.7Mi        +39.26%

                            │     Old      │                  New                  │
                            │     B/op     │     B/op      vs base                 │
Marshal/pbName=Action-24      4.371Ki ± 0%   4.387Ki ± 0%  +0.36% (p=0.000 n=10)
Marshal/pbName=Tree-24        14.57Mi ± 0%   14.61Mi ± 0%  +0.27% (p=0.000 n=10)
Marshal/pbName=Command-24     12.32Ki ± 0%   12.32Ki ± 0%       ~ (p=0.087 n=10)
Unmarshal/pbName=Action-24    12.48Ki ± 0%   12.48Ki ± 0%       ~ (p=1.000 n=10) ¹
Unmarshal/pbName=Tree-24      45.04Mi ± 0%   45.04Mi ± 0%       ~ (p=0.677 n=10)
Unmarshal/pbName=Command-24   37.67Ki ± 0%   37.67Ki ± 0%       ~ (p=1.000 n=10)
geomean                       161.0Ki        161.2Ki       +0.11%
¹ all samples are equal

                            │     Old     │                      New                      │
                            │  allocs/op  │   allocs/op     vs base                       │
Marshal/pbName=Action-24       1.000 ± 0%       2.000 ± 0%      +100.00% (p=0.000 n=10)
Marshal/pbName=Tree-24         1.000 ± 0%   10589.000 ± 0%  +1058800.00% (p=0.000 n=10)
Marshal/pbName=Command-24      1.000 ± 0%       1.000 ± 0%             ~ (p=1.000 n=10) ¹
Unmarshal/pbName=Action-24     220.0 ± 0%       220.0 ± 0%             ~ (p=1.000 n=10) ¹
Unmarshal/pbName=Tree-24      882.1k ± 0%      882.1k ± 0%             ~ (p=1.000 n=10) ¹
Unmarshal/pbName=Command-24    644.0 ± 0%       644.0 ± 0%             ~ (p=1.000 n=10) ¹
geomean                        70.71            371.9           +425.99%
¹ all samples are equal
```